refactor: move sync.go to syncQueue pkg for all banckend services
