+++ services/apiserver/services/agent_internal.go
@@ -3,6 +3,7 @@ package services
 import (
 	"github.com/aenjoy/iot-lubricant/services/corepkg/datastore"
 	"github.com/aenjoy/iot-lubricant/services/corepkg/repo"
+	"github.com/aenjoy/iot-lubricant/services/corepkg/syncQueue"
 )
 
 var _ IAgentService = (*AgentService)(nil)
@@ -10,5 +11,5 @@ var _ IAgentService = (*AgentService)(nil)
 type AgentService struct {
 	db    repo.ICoreDb
 	store *datastore.DataStore
-	*SyncTaskQueue
+	*syncQueue.SyncTaskQueue
 }
