+++ services/apiserver/services/backend/ioc.go
@@ -1,9 +1,9 @@
 package backend
 
 import (
-	"github.com/aenjoy/iot-lubricant/services/apiserver/services"
 	"github.com/aenjoy/iot-lubricant/services/corepkg/datastore"
 	"github.com/aenjoy/iot-lubricant/services/corepkg/ioc"
+	"github.com/aenjoy/iot-lubricant/services/corepkg/syncQueue"
 )
 
 var (
@@ -20,7 +20,7 @@ func (ReportHandler) Weight() uint16 {
 func (r *ReportHandler) Init() error {
 	//r.AgentService = ioc.Controller.Get(ioc.APP_NAME_CORE_GATEWAY_AGENT_SERVICE).(*services.AgentService)
 	r.dataCli = ioc.Controller.Get(ioc.APP_NAME_CORE_DATABASE_STORE).(*datastore.DataStore)
-	r.SyncTaskQueue = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_SyncTask_SERVICE).(*services.SyncTaskQueue)
+	r.SyncTaskQueue = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_SyncTask_SERVICE).(*syncQueue.SyncTaskQueue)
 	go r.handler()
 	return nil
 }
