+++ services/apiserver/services/ioc.go
@@ -5,18 +5,19 @@ import (
 	"github.com/aenjoy/iot-lubricant/services/corepkg/datastore"
 	"github.com/aenjoy/iot-lubricant/services/corepkg/ioc"
 	"github.com/aenjoy/iot-lubricant/services/corepkg/repo"
+	"github.com/aenjoy/iot-lubricant/services/corepkg/syncQueue"
 )
 
 var (
 	_ ioc.Object = (*AgentService)(nil)
 	_ ioc.Object = (*GatewayService)(nil)
-	_ ioc.Object = (*SyncTaskQueue)(nil)
+	_ ioc.Object = (*syncQueue.SyncTaskQueue)(nil)
 )
 
 func (s *GatewayService) Init() error {
 	s.db = ioc.Controller.Get(ioc.APP_NAME_CORE_DATABASE).(*repo.CoreDb)
 	s.store = ioc.Controller.Get(ioc.APP_NAME_CORE_DATABASE_STORE).(*datastore.DataStore)
-	s.SyncTaskQueue = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_SyncTask_SERVICE).(*SyncTaskQueue)
+	s.SyncTaskQueue = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_SyncTask_SERVICE).(*syncQueue.SyncTaskQueue)
 	return nil
 }
 
@@ -31,7 +32,7 @@ func (s *GatewayService) Version() string {
 func (a *AgentService) Init() error {
 	a.db = ioc.Controller.Get(ioc.APP_NAME_CORE_DATABASE).(*repo.CoreDb)
 	a.store = ioc.Controller.Get(ioc.APP_NAME_CORE_DATABASE_STORE).(*datastore.DataStore)
-	a.SyncTaskQueue = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_SyncTask_SERVICE).(*SyncTaskQueue)
+	a.SyncTaskQueue = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_SyncTask_SERVICE).(*syncQueue.SyncTaskQueue)
 	return nil
 }
 
@@ -43,15 +44,15 @@ func (*AgentService) Version() string {
 	return ""
 }
 
-func (s *SyncTaskQueue) Init() error {
+func (s *syncQueue.SyncTaskQueue) Init() error {
 	s.Mq = ioc.Controller.Get(ioc.APP_NAME_CORE_Internal_MQ_SERVICE).(mq.Mq)
 	return nil
 }
 
-func (SyncTaskQueue) Weight() uint16 {
+func (syncQueue.SyncTaskQueue) Weight() uint16 {
 	return ioc.CoreSyncTaskSystem
 }
 
-func (SyncTaskQueue) Version() string {
+func (syncQueue.SyncTaskQueue) Version() string {
 	return ""
 }
